Add JWT auth and users

+++ backend/controllers/goalController.js
@@ -1,13 +1,14 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalsModel");
+const User = require("../models/userModel");
 
 // @desc    Get goals (for user)
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
   //we are not using try/catch and instead using express-async-handler
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id});
 
   res.status(200).json(goals);
 });
@@ -21,11 +22,12 @@ const setGoal = asyncHandler(async (req, res) => {
     throw new Error("Please add a text field");
   }
 
-  const goal = await Goal.create({
-    text: req.body.text,
-  });
+  const text = req.body.text;
 
-  console.log(req.body);
+  const goal = await Goal.create({
+    user: req.user.id,
+    text: text,
+  });
 
   res.status(200).json(goal);
 });
@@ -39,6 +41,20 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id)
+
+  // Check for user
+  if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Make sure the logged in user matches the goal user
+  if(goal.user.toString() != user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -57,6 +73,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id)
+
+  // Check for user
+  if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Make sure the logged in user matches the goal user
+  if(goal.user.toString() != user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await Goal.findByIdAndRemove(req.params.id);
